tweak(*): Tidy up remaining uneeded files now .NET package works
I wanted to keep my changes here as minimal as I could, so I've removed the .NET example as we have that in the repo for that project.

While tidying up I found residual Kiro.dev files that I removed. While removing them I noticed that I should expand the readme section to include more details on how kiro helped and to tweak the wording there.

My goal is to be transparent about the AI use. So I hope that helps. No idea if anyone reads these.